fix(widgetbook): split intervention catalog to satisfy NCL gate (#4267)
Extract OVL50001 intervention Widgetbook stories from catalog_panels.dart
into catalog_panels_intervention.dart so catalog_panels stays under the
1000 non-comment-line repo limit.
